initial variant modal
