+++ src/components/atoms/tooltip/index.tsx
@@ -1,3 +1,4 @@
+import React from "react"
 import * as RadixTooltip from "@radix-ui/react-tooltip"
 import clsx from "clsx"
