+++ src/components/molecules/info-tooltip/index.tsx
@@ -2,12 +2,10 @@ import React from "react"
 import Tooltip from "../../atoms/tooltip"
 import InfoIcon from "../../fundamentals/icons/info-icon"
 
-const InfoTooltip = ({ content, ...props }) => {
-  return (
-    <Tooltip content={content} {...props}>
-      <InfoIcon className="flex text-grey-40" />
-    </Tooltip>
-  )
-}
+const InfoTooltip = ({ content, ...props }) => (
+  <Tooltip content={content} {...props}>
+    <InfoIcon className="flex text-grey-40" />
+  </Tooltip>
+)
 
 export default InfoTooltip
